requirements/prod.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
defusedxml==0.7.1
Django==4.0.6
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-embed-video==1.4.4
fpdf2==2.5.5
gunicorn==20.1.0
idna==3.3
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.10
whitenoise==6.2.0
XlsxWriter==3.0.3
